third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/ownerDocument-import-expected.txt

HTMLTemplateElement content ownerDocument in HTML imports should differ from import document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS template.ownerDocument is not template.content.ownerDocument
PASS successfullyParsed is true

TEST COMPLETE
